refactor(GameRecordMaker): 현재 다리 횡단 상태를 이전 횡단 상태에 추가하는 기능 수정
상수 사용을 피하기 위해 Iterator를 사용하여 추가하도록 변경
