+++ src/main/java/bridge/domain/model/GameRecordGenerator.java
